+++ src/types/utility-types/late.ts
@@ -52,6 +52,8 @@ export class Late<S, T> extends Type<S, T> {
 
 export type ILateType<S, T> = () => IType<S, T>
 
+export function late<S = any, T = any>(type: ILateType<S, T>): IType<S, T>
+export function late<S = any, T = any>(name: string, type: ILateType<S, T>): IType<S, T>
 /**
  * Defines a type that gets implemented later. This is usefull when you have to deal with circular dependencies.
  * Please notice that when defining circular dependencies TypeScript is'nt smart enought to inference them.
@@ -69,14 +71,13 @@ export type ILateType<S, T> = () => IType<S, T>
  * ```
  * 
  * @export
+ * @alias types.late
  * @template S
  * @template T
  * @param {string} [name] The name to use for the type that will be returned.
  * @param {ILateType<S, T>} type A function that returns the type that will be defined.
  * @returns {IType<S, T>} 
  */
-export function late<S = any, T = any>(type: ILateType<S, T>): IType<S, T>
-export function late<S = any, T = any>(name: string, type: ILateType<S, T>): IType<S, T>
 export function late<S, T>(nameOrType: any, maybeType?: ILateType<S, T>): IType<S, T> {
     const name = typeof nameOrType === "string" ? nameOrType : `late(${nameOrType.toString()})`
     const type = typeof nameOrType === "string" ? maybeType : nameOrType