+++ src/types/utility-types/maybe.ts
@@ -11,6 +11,7 @@ const nullType = optional(literal(null), null)
  * Maybe will make a type nullable, and also null by default. 
  * 
  * @export
+ * @alias types.maybe
  * @template S 
  * @template T 
  * @param {IType<S, T>} type The type to make nullable