+++ src/core/mst-operations.ts
@@ -158,7 +158,8 @@ export function revertPatch(target: IStateTreeNode, patch: IJsonPatch | IJsonPat
 }
 
 export interface IPatchRecorder {
-    patches: IJsonPatch[]
+    patches: ReadonlyArray<IJsonPatch>
+    cleanPatches: ReadonlyArray<IJsonPatch>
     stop(): any
     replay(target: IStateTreeNode): any
     undo(): void
@@ -172,6 +173,8 @@ export interface IPatchRecorder {
  * export interface IPatchRecorder {
  *      // the recorded patches
  *      patches: IJsonPatch[]
+ *      // the same set of recorded patches, but without undo information, making them smaller and compliant with json-patch spec
+ *      cleanPatches: IJSonPatch[]
  *      // stop recording patches
  *      stop(): any
  *      // apply all the recorded patches on the given object
@@ -189,6 +192,9 @@ export interface IPatchRecorder {
 export function recordPatches(subject: IStateTreeNode): IPatchRecorder {
     let recorder = {
         patches: [] as IJsonPatch[],
+        get cleanPatches() {
+            return this.patches.map(stripPatch)
+        },
         stop() {
             disposer()
         },
@@ -230,7 +236,7 @@ export function applyAction(
 }
 
 export interface IActionRecorder {
-    actions: ISerializedActionCall[]
+    actions: ReadonlyArray<ISerializedActionCall>
     stop(): any
     replay(target: IStateTreeNode): any
 }
@@ -556,7 +562,7 @@ import {
 } from "./action"
 import { runInAction, IObservableArray, ObservableMap } from "mobx"
 import { Node, getStateTreeNode, IStateTreeNode, isStateTreeNode } from "./node"
-import { IJsonPatch, splitJsonPath, invertPatch } from "./json-patch"
+import { IJsonPatch, splitJsonPath, invertPatch, stripPatch } from "./json-patch"
 import { IDisposer, fail, asArray } from "../utils"
 import { ISnapshottable, IType } from "../types/type"
 import { isType } from "../types/type-flags"