+++ changelog.md
@@ -1,7 +1,10 @@
 
-# Next
+# 0.9.2
 
-`revertPatch`, `patchRecorder.cleanPatches`, `patchRecorder.undo`, `onPatch(target, handler, includeUndoInfo)`
+Introduced the concept of reverse patches, see [#231](https://github.com/mobxjs/mobx-state-tree/pull/231/)
+* Introduced the `revertPatch` operation, that takes a patch or list of patches, and reverse applies it to the target.
+* `onPatch` now takes a second argument, `includeOldValue`, defaulting to `false`, which, if set to true, includes in the patch any value that is being overwritten as result of the patch. Setting this option to true produces patches that can be used with `revertPatch`
+* `patchRecorder` now introduces additional fields / methods to be able to reverse apply changes: `patchRecorder.cleanPatches`, `patchRecorder.undo`
 
 # 0.9.1